+++ CHANGELOG.md
@@ -1,6 +1,10 @@
-0.1.0 / 2016-08-26
+0.1.4 / 2016-09-01
 ------------------
-- Added HTMLElementWrapp.innerHTML
+- Added HTMLElementWrapper.enable and disable.
+
+0.1.3 / 2016-08-26
+------------------
+- Added HTMLElementWrapper.innerHTML
 
 0.1.0 / 2016-08-26
 ------------------